pyproject.toml
[tool.poetry]
name = "corevist-demo-project"
version = "0.1.0"
description = ""
authors = ["Anhelina Babareka <anhielina.babareka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = ">=7.4.2,<7.5.0"
selene = ">=2.0.0rc4,<2.1.0"
allure-pytest = ">=2.13.2,<2.14.0"
pydantic-settings = ">=2.0.3,<2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
